+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
 	"key/aziot-keyd",
 	"key/aziot-key-openssl-engine",
 	"key/aziot-key-openssl-engine-shared",
+	"key/aziot-key-openssl-engine-shared-test",
 	"key/aziot-keys",
 	"key/aziot-keys-common",
 
@@ -35,9 +36,7 @@ members = [
 	"openssl-sys2",
 
 	"pkcs11/pkcs11",
-	"pkcs11/pkcs11-openssl-engine",
 	"pkcs11/pkcs11-sys",
-	"pkcs11/pkcs11-test",
 
 	"tpm/aziot-tpm-client-async",
 	"tpm/aziot-tpm-common-http",
