+++ docs/dev/running/pkcs11/cryptoauthlib.md
@@ -120,4 +120,4 @@ The library should now be configured completely.
 
 1. `pkcs11-tool --module /usr/lib/libcryptoauth.so -IOT` for some reason does not see any attributes of the PKCS#11 objects. Use `p11tool --list-all 'pkcs11:model=ATECC608A'` instead.
 
-1. As mentioned above, the library creates files named like `<slot number>.<actual device slot number>.conf` to store PKCS#11 metadata about dynamically-generated keys. Therefore, deleting these object files will make the PKCS#11 library "forget" about those objects, even if the actual hardware slot is occupied, so it's a handy way to "reset" the device for tests. Note that the library maintains the state of which slots are used in memory as well, so deleting the files only has an effect if the process using the library (`pkcs11-test`, `aziot-keyd`, etc) is restarted.
+1. As mentioned above, the library creates files named like `<slot number>.<actual device slot number>.conf` to store PKCS#11 metadata about dynamically-generated keys. Therefore, deleting these object files will make the PKCS#11 library "forget" about those objects, even if the actual hardware slot is occupied, so it's a handy way to "reset" the device for tests. Note that the library maintains the state of which slots are used in memory as well, so deleting the files only has an effect if the process using the library (`aziot-keyd`, etc) is restarted.
