+++ docs/dev/running/pkcs11/index.md
@@ -118,151 +118,12 @@ sudo mkdir -p /etc/pkcs11/modules
 
 However, `--provider` is still useful to filter out tokens from other PKCS#11 libraries.
 
+
 ---
 
 
-## `pkcs11-test`
+## Testing with aziot-keyd
 
-To verify the hardware and PKCS#11 library, there is a standalone `pkcs11-test` binary in this repository that can be used to run some PKCS#11-focused tests.
+To verify your HSM and its PKCS#11 library with aziot-keyd, see [Testing the openssl engine.](../aziot-keyd.md#testing-the-openssl-engine)
 
-1. Build.
-
-    ```sh
-    make pkcs11-test
-    ```
-
-1. Define three variables for the labels of the three certs that will be generated as part of the test.
-
-    ```sh
-    LABEL_1='CA' # A CA cert
-    LABEL_2='Server' # A server cert signed by the CA cert
-    LABEL_3='Client' # A client cert signed by the CA cert
-    ```
-
-1. Define three variables for the algorithms of the private keys of the certs. Valid values are `rsa-2048`, `rsa-4096` and `ec-p256`.
-
-    - `softhsm2`
-
-        ```sh
-        KEY_1_TYPE='rsa-4096'
-        KEY_2_TYPE='rsa-2048'
-        KEY_3_TYPE='rsa-2048'
-
-        # KEY_1_TYPE='rsa-2048'
-        # KEY_2_TYPE='rsa-2048'
-        # KEY_3_TYPE='rsa-2048'
-
-        # KEY_1_TYPE='ec-p256'
-        # KEY_2_TYPE='ec-p256'
-        # KEY_3_TYPE='ec-p256'
-        ```
-
-    - `cryptoauthlib`
-
-        ```sh
-        KEY_1_TYPE='ec-p256'
-        KEY_2_TYPE='ec-p256'
-        KEY_3_TYPE='ec-p256'
-        ```
-
-        `cryptoauthlib` only supports `ec-p256` keys.
-
-    - `tpm2-pkcs11`
-
-        ```sh
-        KEY_1_TYPE='rsa-2048'
-        KEY_2_TYPE='rsa-2048'
-        KEY_3_TYPE='rsa-2048'
-
-        # KEY_1_TYPE='ec-p256'
-        # KEY_2_TYPE='ec-p256'
-        # KEY_3_TYPE='ec-p256'
-        ```
-
-        `tpm2-pkcs11` only supports `rsa-2048` and `ec-p256` keys.
-
-1. Run `pkcs11-test`
-
-    ```sh
-    # The `PKCS11_SPY_PATH` env var can be defined to intercept
-    # all PKCS#11 API calls to the PKCS#11 library and print
-    # their inputs and outputs.
-    #
-    # export PKCS11_SPY_PATH='/usr/lib64/pkcs11/pkcs11-spy.so'
-    # export PKCS11_SPY_PATH='/usr/lib/arm-linux-gnueabihf/pkcs11/pkcs11-spy.so'
-
-    PKCS11_TEST_PATH="$PWD/target/x86_64-unknown-linux-gnu/debug/pkcs11-test"
-    # PKCS11_TEST_PATH="$PWD/target/armv7-unknown-linux-gnueabhif/debug/pkcs11-test"
-
-    # Generate three asymmetric keys to be used as the private keys
-    # for the three certs
-    "$PKCS11_TEST_PATH" generate-key-pair \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_1$PIN_SUFFIX" \
-        --type "$KEY_1_TYPE" &&
-    "$PKCS11_TEST_PATH" generate-key-pair \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_2$PIN_SUFFIX" \
-        --type "$KEY_2_TYPE" &&
-    "$PKCS11_TEST_PATH" generate-key-pair \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_3$PIN_SUFFIX" \
-        --type "$KEY_3_TYPE" &&
-
-    # Load them and print them
-    "$PKCS11_TEST_PATH" load --keys \
-        "pkcs11:$TOKEN_PARAM;object=$LABEL_1" \
-        "pkcs11:$TOKEN_PARAM;object=$LABEL_2" \
-        "pkcs11:$TOKEN_PARAM;object=$LABEL_3" &&
-
-    # Generate the CA cert
-    "$PKCS11_TEST_PATH" generate-ca-cert \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_1$PIN_SUFFIX" \
-        --subject 'CA Inc' \
-        --out-file "$PWD/ca.pem" &&
-
-    # Generate the server cert
-    "$PKCS11_TEST_PATH" generate-server-cert \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_2$PIN_SUFFIX" \
-        --subject 'Server LLC' \
-        --ca-cert "$PWD/ca.pem" \
-        --ca-key "pkcs11:$TOKEN_PARAM;object=$LABEL_1$PIN_SUFFIX" \
-        --out-file "$PWD/server.pem" &&
-
-    # Generate the client cert
-    "$PKCS11_TEST_PATH" generate-client-cert \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_3$PIN_SUFFIX" \
-        --subject 'Client GmbH' \
-        --ca-cert "$PWD/ca.pem" \
-        --ca-key "pkcs11:$TOKEN_PARAM;object=$LABEL_1$PIN_SUFFIX" \
-        --out-file "$PWD/client.pem" &&
-
-    # Start a hello-world web server that serves HTTPS on port 8443
-    # using the server cert. The server will remain running.
-    "$PKCS11_TEST_PATH" web-server \
-        --cert "$PWD/server.pem" \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_2$PIN_SUFFIX"
-
-    # In another shell, connect to the server in various ways
-    # to verify that TLS works.
-    #
-    # 1. `openssl s_client`
-    < /dev/null openssl s_client -connect 127.0.0.1:8443 &&
-
-    # 2. `curl`
-    #
-    #    This should receive an HTTP 200 response with the body
-    #    "Hello, world!". `-k` is needed since the CA cert is not in
-    #    the OS trusted roots.
-    curl -kD - https://127.0.0.1:8443 &&
-
-    # 3. `pkcs11-test web-client`. This uses the client cert for
-    #    TLS client authentication. It should receive
-    #    an HTTP 200 response with the body "Hello, world!"
-    #
-    #    Both the client and the server will also print the cert chain
-    #    of the other, and assert that the other's cert is signed by
-    #    the common CA cert.
-    "$PKCS11_TEST_PATH" web-client \
-        --cert "$PWD/client.pem" \
-        --key "pkcs11:$TOKEN_PARAM;object=$LABEL_3$PIN_SUFFIX"
-    ```
-
-If the three client tests pass successfully, ie `openssl s_client` completes the handshake successfully and prints the server cert details, and `curl` and `pkcs11-test web-client` print the hello-world response from the server, then the hardware and PKCS#11 library ought to be suitable for `aziot-keyd` to use. If there are errors such as crashes or signature verification failures, then it might be a problem with the hardware configuration, a bug in the PKCS#11 library, or a bug in the PKCS#11-related code in this repository.
+If the script mentioned there completes successfully, then the hardware and PKCS#11 library ought to be suitable for `aziot-keyd` to use. If there are errors such as crashes or signature verification failures, then it might be a problem with the hardware configuration, a bug in the PKCS#11 library, or a bug in the PKCS#11-related code in this repository.
