+++ key/aziot-key-openssl-engine-shared-test/src/tokio_openssl2.rs
@@ -1,87 +1,5 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-pub(crate) async fn connect(
-    stream: std::net::TcpStream,
-    cert_chain_path: &std::path::Path,
-    private_key: &openssl::pkey::PKey<openssl::pkey::Private>,
-    domain: &str,
-) -> std::io::Result<bytes::BytesMut> {
-    use futures_util::StreamExt;
-
-    let stream = tokio::net::TcpStream::from_std(stream)?;
-
-    let mut tls_connector = openssl::ssl::SslConnector::builder(openssl::ssl::SslMethod::tls())?;
-    tls_connector.set_certificate_chain_file(cert_chain_path)?;
-    tls_connector.set_private_key(private_key)?;
-
-    // The root of the client cert is the CA, and we expect the server cert to be signed by this same CA.
-    // So add it to the cert store.
-    let ca_cert = {
-        let cert_chain_file = std::fs::read(cert_chain_path)?;
-        let mut cert_chain = openssl::x509::X509::stack_from_pem(&cert_chain_file)?;
-        cert_chain.pop().unwrap()
-    };
-    tls_connector.cert_store_mut().add_cert(ca_cert)?;
-
-    // Log the server cert chain. Does not change the verification result from what openssl already concluded.
-    tls_connector.set_verify_callback(
-        openssl::ssl::SslVerifyMode::PEER,
-        |openssl_verification_result, context| {
-            println!("Server cert:");
-            let chain = context.chain().unwrap();
-            for (i, cert) in chain.into_iter().enumerate() {
-                println!(
-                    "    #{}: {}",
-                    i + 1,
-                    cert.subject_name()
-                        .entries()
-                        .next()
-                        .unwrap()
-                        .data()
-                        .as_utf8()
-                        .unwrap()
-                );
-            }
-            println!(
-                "openssl verification result: {}",
-                openssl_verification_result
-            );
-            openssl_verification_result
-        },
-    );
-
-    let tls_connector = tls_connector.build();
-
-    let stream = tokio_openssl::connect(tls_connector.configure()?, domain, stream)
-        .await
-        .map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
-
-    let (mut send_request, connection) = hyper::client::conn::handshake(stream)
-        .await
-        .map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
-
-    let mut request = hyper::Request::new(Default::default());
-    *request.uri_mut() = hyper::Uri::from_static("/");
-    let send_request = send_request.send_request(request);
-
-    let connection = connection.without_shutdown();
-    let _ = tokio::spawn(connection);
-
-    let response = send_request
-        .await
-        .map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
-    let mut response_body = response.into_body();
-
-    let mut response = bytes::BytesMut::new();
-
-    while let Some(chunk) = response_body.next().await {
-        let chunk = chunk.map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
-        response.extend_from_slice(&chunk);
-    }
-
-    Ok(response)
-}
-
 type HandshakeFuture = std::pin::Pin<
     Box<
         dyn std::future::Future<
