+++ openssl-build/src/lib.rs
@@ -4,7 +4,7 @@
 #![warn(clippy::all, clippy::pedantic)]
 #![allow(clippy::must_use_candidate)]
 
-/// Emits `ossl110` and `ossl111` cfgs based on the version of openssl.
+/// Emits `ossl110` cfg based on the version of openssl.
 pub fn define_version_number_cfg() {
     let openssl_version = std::env::var("DEP_OPENSSL_VERSION_NUMBER")
         .expect("DEP_OPENSSL_VERSION_NUMBER must have been set by openssl-sys");
@@ -14,10 +14,6 @@ pub fn define_version_number_cfg() {
     {
         if openssl_version >= 0x01_01_00_00_0 {
             println!("cargo:rustc-cfg=ossl110");
-        }
-
-        if openssl_version >= 0x01_01_01_00_0 {
-            println!("cargo:rustc-cfg=ossl111");
         }
     }
 }
