--- pkcs11/pkcs11-openssl-engine/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "pkcs11-openssl-engine"
-version = "0.1.0"
-license = "MIT"
-authors = ["Azure IoT Edge Devs"]
-edition = "2018"
-build = "build/main.rs"
-
-
-[dependencies]
-foreign-types-shared = "0.1"
-log = "0.4"
-openssl = "0.10"
-openssl-errors = "0.1"
-openssl-sys = "0.9"
-
-openssl2 = { path = "../../openssl2/" }
-openssl-sys2 = { path = "../../openssl-sys2/" }
-pkcs11 = { path = "../pkcs11/" }
-pkcs11-sys = { path = "../pkcs11-sys/" }
-
-
-[build-dependencies]
-openssl-build = { path = "../../openssl-build/" }
