--- pkcs11/pkcs11-test/build.rs
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-
-#![deny(rust_2018_idioms)]
-#![warn(clippy::all, clippy::pedantic)]
-
-fn main() {
-    openssl_build::define_version_number_cfg();
-}
